Revert 10.5.1 changes; those will go in separate PR.
This reverts commits 42f33d04f7869e611bae976e5613384e0268caf5 and 57ddf10cd09db2f751f226a6cceaab5d839bdbfa.
