+++ classes/OpenXdmod/Migration/Version1050To1100/DatabasesMigration.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * Update database from version 10.5.1 to 11.0.0
+ * Update database from version 10.5.0 to 11.0.0
  */
 
-namespace OpenXdmod\Migration\Version1051To1100;
+namespace OpenXdmod\Migration\Version1050To1100;
 
 use OpenXdmod\Migration\DatabasesMigration as AbstractDatabasesMigration;
 use CCR\DB;
@@ -11,7 +11,7 @@ use CCR\DB\MySQLHelper;
 use ETL\Utilities;
 
 /**
- * Migrate databases from version 10.5.1 to 11.0.0
+ * Migrate databases from version 10.5.0 to 11.0.0
  */
 class DatabasesMigration extends AbstractDatabasesMigration
 {
