+++ docs/_config.yml
@@ -36,7 +36,7 @@ defaults:
       type: "pages"
     values:
       layout: "page"
-      prev_sw_version: "11.0.0"
+      prev_sw_version: "10.5.0"
       sw_version: "11.0.1"
       style: "effervescence"
       tocversion: "v11_0toc"
